+++ src/vsmtp/vsmtp-common/src/state.rs
@@ -66,3 +66,11 @@ impl Default for StateSMTP {
         Self::Connect
     }
 }
+
+impl StateSMTP {
+    /// As the email been received at the current stage ?
+    #[must_use]
+    pub const fn email_received(&self) -> bool {
+        matches!(self, Self::PostQ | Self::Delivery)
+    }
+}