+++ src/vsmtp/vsmtp-common/src/status.rs
@@ -15,9 +15,9 @@
  *
 */
 
-use crate::ReplyOrCodeID;
+use crate::{transfer::SmtpConnection, ReplyOrCodeID};
 
-/// Status of the mail context treated by the rule engine
+/// Status of the mail context treated by the rule engine.
 #[derive(Debug, Clone, PartialEq, Eq, strum::AsRefStr, serde::Deserialize, serde::Serialize)]
 #[strum(serialize_all = "snake_case")]
 pub enum Status {
@@ -44,14 +44,13 @@ pub enum Status {
     /// used to send data from .vsl to vsmtp's server
     Packet(String),
 
-    // TODO: add rule / action name,
     /// the email as been delegated to another service.
-    Delegated,
+    #[serde(skip)]
+    Delegated(SmtpConnection),
 
-    /// the rule engine must skip all rules until the given
-    /// rule because the message received is a delegation
-    /// result, it would be worthless to re-execute all rules.
-    DelegationResult(String),
+    /// the rule engine must skip all rules until a given
+    /// rule received in the email's header.
+    DelegationResult,
 }
 
 impl Status {
@@ -60,7 +59,7 @@ impl Status {
     #[must_use]
     pub const fn stop(&self) -> bool {
         match self {
-            Status::Faccept(_) | Status::Deny(_) | Status::Quarantine(_) | Status::Delegated => {
+            Status::Faccept(_) | Status::Deny(_) | Status::Quarantine(_) | Status::Delegated(_) => {
                 true
             }
             _ => false,