+++ src/vsmtp/vsmtp-server/src/delivery/deferred.rs
@@ -37,6 +37,7 @@ pub async fn flush_deferred_queue(
     for path in dir_entries {
         let process_message = ProcessMessage {
             message_id: path?.path().file_name().unwrap().to_string_lossy().into(),
+            delegated: false,
         };
 
         if let Err(e) =
@@ -174,6 +175,7 @@ mod tests {
             std::sync::Arc::new(resolvers),
             ProcessMessage {
                 message_id: "test_deferred".to_string(),
+                delegated: false,
             },
         )
         .await